producer/requirements.txt
uvicorn
fastapi
pymongo
confluent_kafka
python-dotenv
